### 2025-06-06 #8361 @xoolive
Hello @ramondalmau @cathaypacific8747 @sfo @niclaswue @krumjan: would you be willing to participate in the review of this software paper? The reviewing process is fully open, looks like a checklist in a GitHub issue, and is focused on both the software and the paper. You can check in the previous entry links to reviews for other similar paper, including the one for the traffic library.

co-authored with @junzis @emctoo @michelkhalaf
