### 2025-06-12 1fb76a6
implement wind prediction using MeteoFrance Arpege grib files
