+++ src/tangram/plugins/system/readme.md
@@ -1,0 +1,3 @@
+# System plugin
+
+This plugin does not provide any API endpoints, but it is used to start an async task that will run in the background. This task is used to perform various system-related operations, such as checking the status of the system, performing maintenance tasks, etc.
