### 2025-06-13 3d634dc
document the wind field plugin
