+++ docs/plugins/index.md
@@ -9,6 +9,6 @@ The tangram platform can be extended with plugins to enhance its functionality. 
 Examples of plugins:
 
 - [Map data receivers](examples/sensors.md): display the sensors' positions on the map
+- [Wind fields](examples/windfield.md): display the wind field on the map using weather prediction data
 - [Origin and destination city pair](examples/citypair.md): display the origin and destination airports of flights
-- [Wind fields](examples/windfield.md): display the wind field on the map using the fastmeteo data
-- [Contrails](examples/contrails.md): display the contrails on the map using the fastmeteo data
+- [Contrails](examples/contrails.md): display the contrails on the map using weather prediction data
