+++ docs/plugins/examples/windfield.md
@@ -1,1 +1,232 @@
-# Wind field with fastmeteo
+# Add a wind field layer
+
+## Statement of need
+
+Meteorological data is essential for understanding the impact of weather on aviation operations. Meteo France provides weather prediction data from their [ARPEGE model](https://meteofrance.com/actualites-et-dossiers/modeles-prevision-meteo). The most basic features include zonal and meridional wind components, but also temperature, pressure, and humidity. More advanced features are also available but we will not cover them here.
+
+A [web API](https://portail-api.meteofrance.fr/web/en/api/ARPEGE) is available after registration, but we will prefer here the GRIB files that can also be downloaded directly from the [data.gouv.fr](https://www.data.gouv.fr/fr/datasets/donnees-pnt-retention-14-jours/) initiative, also provided by Meteo France under an Open License.
+
+New predictions are published every 6 hours, and the data is available for 3 days in advance. The data is provided in GRIB format, which is a standard format for meteorological data. In each file, each grid point represents spatio -temporal coordinates, and the data is provided for each hour of the day. The data is available at a resolution of 0.1 degrees, which is approximately 11 km at the equator. Altitude is provided in isobars, which is a standard unit for atmospheric pressure.
+
+## Implementation
+
+The objective of this plugin is to display a wind field on the map, at an isobar specified by the user in a slider. We need to work on the following steps:
+
+1. on the backend side, download the GRIB files from the data.gouv.fr initiative, and provide an API endpoint to fetch the wind field data for a specific isobar
+2. on the frontend side, create a Vue component that will display a slider to select an isobar, and display the corresponding wind field on the map.
+
+### 1. Implement a backend plugin for wind information
+
+Create a `weather` folder in the `src/tangram/plugins/` directory, and register the plugin in the `__init__.py` file. The plugin, associated with the `/weather` router will provide an API endpoint to fetch the wind field data for a specific isobar.
+
+The logic for downloading the data is implemented in the `arpege.py` file.
+
+The code is self-explanatory, but the following points are worth noting:
+
+- in `__init__.py`, the `wind(isobaric: int)` function returns a `ORJSONResponse` (from `fastapi.responses`) rather than a `JSONResponse`. This is because the data is large and we want to avoid the overhead of converting it to JSON. The `ORJSONResponse` is a faster alternative, based on [**`orjson`**](https://github.com/ijl/orjson) that can handle large data efficiently. It also automatically deals with datetime objects, which are used in the GRIB data.
+
+!!! tip
+
+    Extra Python libraries can be included with the `uv add` command.
+
+    The command edits both the `pyproject.toml` and the `uv.lock` files:
+
+    ```shell
+    uv add xarray cfgrib orjson
+    ```
+
+- in `arpege.py`, the download of the GRIB file is wrapped in a try/except block. If the file is not found (or Internet is momentarily unavailable), the function will try to get the previous file (6 hours before) and return it. It can be useful to fallback to older files when the newer versions are not yet available.
+
+- the `xarray` library offers a `load_dataset` and an `open_dataset` function to read GRIB files. The `open_dataset` function is used here, as it allows to read the file without loading it entirely into memory, which is useful for large files.
+
+!!! warning
+
+    There is an issue with the packaging of the `ecCodes` library that is used by `xarray` to read GRIB files. The specific version of the `ecCodes` library for Linux/arm64 is not yet available at the time of writing this documentation.
+
+    **You are most likely to be impacted by this issue if you run this plugin on Apple Silicon Mac computers.**
[…]
