### 2025-06-16 1f732ca
city pair plugin implementation and documentation
