+++ docs/plugins/examples/citypair.md
@@ -1,3 +1,165 @@
 # Origin and destination city pair
 
-(hatching)
+## Statement of need
+
+The city pair plugin displays the origin and destination information for the selected aircraft, showing both the airport codes (ICAO) and the city names. This makes it easy to quickly understand the flight route without leaving the main map interface.
+
+## Implementation
+
+The implementation of the city pair plugin is a Vue.js component that displays the origin and destination of a flight when an aircraft is selected on the map. The implementation is located in the `src/plugins/CityPair.vue` file.
+
+!!! tip
+
+    When a Vue component is a plugin, its implementation can be overridden by a custom implementation located in a different directory. This is useful if you want to have several possible implementations for a functionality, or if you want to use a different implementation for a specific use case.
+    This is particularly useful if you have a custom implementation of the city pair plugin using different sources of data (specific airport or airline development API, EUROCONTROL B2B, etc.) that are not necessarily publicly available, or not without a private token.
+
+    If the environment variable `TANGRAM_WEB_CITYPAIR_PLUGIN` is defined, the plugin will be loaded from the specified path. Otherwise, it will be loaded from the default fallback location. You can specify the path to your custom implementation in the `.env` file.
+
+In this example, the flight route information is obtained from the OpenSky Network's route service, which provides origin-destination pairs based on the flight's callsign.
+The city names are retrieved using the `airport_information` function from the [`rs1090-wasm` library](https://www.npmjs.com/package/rs1090-wasm).
+
+### 1. Declare the plugin in the `vite.config.js` file
+
+In order to use the `CityPair` component as a plugin, you need to declare it in the `vite.config.js` file:
+
+```javascript
+plugins: [
+  // ..., other settings
+  dynamicComponentsPlugin({
+    envPath: "../.env",
+    fallbackDir: "/src/plugins/",
+    availablePlugins: [
+      "airportSearch",
+      "systemInfo",
+      "sensorsInfo",
+      // "windField",
+      "cityPair",  // <-- new line
+    ],
+  }),
+],
+```
+
+### 2. Configure proxy for the route API
+
+The city pair plugin needs to fetch route information from an external API.
+
+When we fetch data from an external API, we often encounter CORS (Cross-Origin Resource Sharing) issues, which is a security feature implemented by web browsers to prevent malicious websites from making requests to other domains. This can block our requests to the OpenSky Network's route service.
+To avoid these issues, we can set up a proxy in the `vite.config.js` file to redirect requests to the external API: we create the `/route` endpoint in the proxy configuration, which will redirect requests to the OpenSky Network's route service.
+
+```javascript
+server: {
+  proxy: {
+    // ..., other settings
+    "/route": {
+      target: "https://flightroutes.opensky-network.org",
+      agent: proxy_agent,
+      changeOrigin: true,
+      secure: true,
+      rewrite: () => "/api/routeset",
+    },
+  },
+},
+```
+
+!!! tip
+
+    If you have proxy settings configured in your environment, the `proxy_agent` will automatically use them. This allows you to work with the OpenSky Network's route service without needing to modify the code.
+
+### 3. Implement the Vue component
+
+The `CityPair.vue` component has three main sections: template, script and style.
+
+The template section displays the origin and destination airports with their respective city names. The template section controls the structure of the DOM elements and how they are displayed in the browser. We use here Vue's conditional rendering to display the city pair information only when both origin and destination are available. We also display loading and error messages based on the component's state.
+
+```vue
+<template>
+  <div id="city_pair" v-if="origin && destination">
+    <div class="airport">
+      <span class="icao">{{ origin }}</span>
[…]
