### 2025-06-16 #8361 @danielskatz
> [@xoolive](https://github.com/xoolive) yes, happy to edit this one, you can go ahead and assign me!

Hi - you can assign yourself via `@editorialbot assign me as editor`
