### 2025-06-20 10964ff
refactoring and handling timeseries in redis
