+++ crates/planes/src/history_redis_ts.rs
@@ -1,1175 +1,0 @@
-mod aircraftdb;
-
-use std::collections::{BTreeMap, HashMap};
-use std::io;
-use std::sync::Arc;
-use std::time::{Duration, SystemTime, UNIX_EPOCH};
-
-use anyhow::{Context, Result};
-use clap::Parser;
-use crossterm::{
-    event::{self, DisableMouseCapture, EnableMouseCapture, Event, KeyCode},
-    execute,
-    terminal::{
-        disable_raw_mode, enable_raw_mode, EnterAlternateScreen, LeaveAlternateScreen,
-    },
-};
-use futures::StreamExt;
-use ratatui::{
-    backend::CrosstermBackend,
-    layout::{Constraint, Direction, Layout},
-    style::{Color, Modifier, Style},
-    text::Span, // Removed Text import
-    widgets::{Block, Borders, Cell, Paragraph, Row, Table},
-    Terminal,
-};
-use redis::{AsyncCommands, RedisResult};
-use rs1090::data::patterns::aircraft_information;
-use serde::{Deserialize, Serialize};
-use tokio::sync::Mutex;
-use tracing::{debug, error, info};
-use tracing_subscriber::EnvFilter;
-
-use crate::aircraftdb::{aircraft, Aircraft};
-
-/// Jet1090 message format
-#[derive(Debug, Clone, Serialize, Deserialize)]
-struct Jet1090Message {
-    icao24: String,
-    df: String,
-    bds: Option<String>,
-    timestamp: f64,
-    callsign: Option<String>,
-    altitude: Option<f64>,
-    latitude: Option<f64>,
-    longitude: Option<f64>,
-    track: Option<f64>,
-    #[serde(flatten)]
-    extra: HashMap<String, serde_json::Value>,
-}
-
-/// State Vector representing an aircraft position and state
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct StateVector {
-    pub icao24: String,
-    pub registration: Option<String>,
-    pub typecode: Option<String>,
-    pub lastseen: f64,
-    pub firstseen: f64,
-    pub callsign: Option<String>,
-    pub latitude: Option<f64>,
-    pub longitude: Option<f64>,
-    pub altitude: Option<f64>,
-    pub track: Option<f64>,
-}
-
-/// Redis-based state vectors collection with Time Series support
-#[derive(Debug)]
-struct RedisTimeSeriesState {
-    aircraft_db: BTreeMap<String, Aircraft>,
-    redis_client: redis::Client,
-    aggregation_interval: u64, // Aggregation interval in seconds
-    retention_msecs: i64,      // Retention period in milliseconds
-    lastwrite_expiry: u64,     // TTL for lastwrite keys
-}
-
-impl RedisTimeSeriesState {
-    async fn new(
-        redis_url: &str, aggregation_interval: u64, history_expiry: u64,
-    ) -> Result<Self> {
-        let redis_client = redis::Client::open(redis_url)
-            .context("Failed to create Redis client for time series state")?;
-
-        let mut conn = redis_client.get_multiplexed_async_connection().await?;
-        let _: RedisResult<String> = conn.ping().await;
-
-        info!("Using data aggregation interval of {} seconds", aggregation_interval);
-        info!("History data will expire after {} seconds", history_expiry);
-
-        Ok(Self {
-            aircraft_db: aircraft().await,
-            redis_client,
-            aggregation_interval,
-            retention_msecs: (history_expiry as i64) * 1000,
-            lastwrite_expiry: 600, // 10 minutes TTL for lastwrite keys
-        })
-    }
-
-    async fn get_aircraft(&self, icao24: &str) -> Result<Option<StateVector>> {
[…]
