+++ crates/planes/Cargo.toml
@@ -33,10 +33,6 @@ ratatui = { version = "0.24.0" }
 crossterm = "0.27.0"
 
 [[bin]]
-name = "history"
-path = "src/history.rs"
-
-[[bin]]
 name = "history_redis"
 path = "src/history_redis.rs"
 
