### 2025-06-25 #8361 @danielskatz
@xoolive - I actually have no formal role here.  I just saw the error earlier and tried to respond.  The track editor @openjournals/sbcs-eics) needs to respond to you. I think there likely is a long queue, but am not sure.
