### 2025-06-25 #8361 @samhforbes
I should add though @xoolive the paper is longer than we would normally accept - we are typically looking for a wordcount around 1000. While reviewers are being looked for perhaps you can reduce this?
