### 2025-06-26 #8361 @xoolive
Thank you @samhforbes, I will adjust the paper then, and move the more verbose content to the documentation website.
