+++ .github/workflows/podman.yml
@@ -26,7 +26,7 @@ jobs:
           podman --version
 
       - name: Cache Podman layers
-        uses: actions/cache@v3
+        uses: actions/cache@v4
         with:
           path: /tmp/.podman-cache
           key: ${{ runner.os }}-podman-${{ github.sha }}
