### 2025-07-01 91a9335
Bump extractions/setup-just from 1 to 3
Bumps [extractions/setup-just](https://github.com/extractions/setup-just) from 1 to 3.
- [Release notes](https://github.com/extractions/setup-just/releases)
- [Commits](https://github.com/extractions/setup-just/compare/v1...v3)

---
updated-dependencies:
- dependency-name: extractions/setup-just
  dependency-version: '3'
  dependency-type: direct:production
  update-type: version-update:semver-major
...

Signed-off-by: dependabot[bot] <support@github.com>
