### 2025-07-01 c8455d1
reduce word count in paper.md
