+++ paper/paper.bib
@@ -3,17 +3,7 @@
   author = {{EUROCONTROL}},
   url    = {https://www.eurocontrol.int/dashboard/rnd-data-archive}
 }
-@article{nilsson_swedish_2023,
-  title   = {Swedish Civil Air Traffic Control Dataset},
-  author  = {Nilsson, Jens and Unger, Jonas},
-  year    = {2023},
-  month   = jun,
-  journal = {Data in Brief},
-  volume  = {48},
-  pages   = {109240},
-  doi     = {10.1016/j.dib.2023.109240},
-  url     = {https://linkinghub.elsevier.com/retrieve/pii/S2352340923003591}
-}
+
 @article{olive_traffic_2019,
   title   = {{traffic}, a Toolbox for Processing and Analysing Air Traffic Data},
   author  = {Olive, Xavier},
@@ -87,17 +77,7 @@
   doi     = {10.1016/j.trc.2021.103119},
   url     = {https://linkinghub.elsevier.com/retrieve/pii/S0968090X21001388}
 }
-@article{zhao_incremental_2021,
-  title   = {An Incremental Clustering Method for Anomaly Detection in Flight Data},
-  author  = {Zhao, Weizun and Li, Lishuai and Alam, Sameer and Wang, Yanjun},
-  year    = {2021},
-  month   = nov,
-  journal = {Transportation Research Part C: Emerging Technologies},
-  volume  = {132},
-  pages   = {103406},
-  doi     = {10.1016/j.trc.2021.103406},
-  url     = {https://linkinghub.elsevier.com/retrieve/pii/S0968090X21004010}
-}
+
 @article{felux_impacts_2024,
   title   = {Impacts of {{Global Navigation Satellite System Jamming}} on {{Aviation}}},
   author  = {Felux, Michael and Fol, Patric and Figuet, Benoit and Waltert, Manuel and Olive, Xavier},
@@ -109,11 +89,3 @@
   doi     = {10.33012/navi.657},
   url     = {https://navi.ion.org/content/71/3/navi.657}
 }
-@inproceedings{vos_transformer_2024,
-  title     = {A {{Transformer-based Trajectory Prediction Model}} to {{Support Air Traffic Demand Forecasting}}},
-  booktitle = {Proceedings of the 11th {{International Conference}} for {{Research}} in {{Air Transportation}}},
-  author    = {Vos, Reinier and Sun, Junzi and Hoekstra, Jacco},
-  year      = {2024},
-  address   = {Singapore}
-}
-
