+++ paper/paper.md
@@ -29,6 +29,7 @@ date: "14 May 2025"
 ---
 
 # Summary
+
 tangram is an open research framework for ADS-B and Mode S flight surveillance data. Initially developed for turbulence detection research, it has since evolved into a versatile tool applicable to various real-time aviation research topics, including GNSS jamming detection, aviation weather monitoring, emission analysis, and airport performance monitoring.
 
 The system comprises a JavaScript-based web application and a backend implemented in Python or Rust, depending on performance requirements. The web application is modular, enabling users to add custom plugins for data processing and analysis. The backend handles data collection, storage, and processing, while the frontend offers a user-friendly interface for visualizing and interacting with the data.
@@ -39,9 +40,9 @@ The whole framework is designed to be extensible, allowing researchers to develo
 
 # Statement of need
 
-The world of aviation involves many stakeholders producing a vast amount of data, but most of it is not publicly available for confidentiality and security reasons. Research in aviation is also often limited by the availability of data, but this has evolved in the past decade thanks to the introduction of open data initiatives [@eurocontrol_rd] [@nilsson_swedish_2023]. ADS-B and Mode S are technologies that allow aircraft to broadcast their position and other flight parameters, making it possible to collect real-time data on air traffic. Such data has been widely used for various research purposes, such as studying flight patterns [@olive_training_2025], estimating aircraft emissions [@seymour_fuel_2020], or monitoring air traffic performance [@schultz_predictive_2021].
+The world of aviation involves many stakeholders producing a vast amount of data, but most of it is not publicly available for confidentiality and security reasons. Research in aviation is also often limited by the availability of data, but this has evolved in the past decade thanks to the introduction of open data initiatives [@eurocontrol_rd].
 
-ADS-B is not encrypted, and it is very easy to build a decoder to receive and decode the data with software-defined radios. This openness has favored the emergence of various open-source projects and tools for working with ADS-B data, such as dump1090 and pyModeS [@sun_pymodes_2020], and platforms like The OpenSky Network [@opensky2014], a non-profit initiative that provides open access to real-time and historical air traffic data, which grants free access to researchers to study various aspects of aviation.
+ADS-B and Mode S enable aircraft to broadcast their position and flight parameters, allowing real-time air traffic data collection. This data supports research on flight patterns [@olive_training_2025], emissions [@seymour_fuel_2020], and air traffic performance [@schultz_predictive_2021]. Since ADS-B is unencrypted and easily decoded with software-defined radios, many open-source tools, such as dump1090, pyModeS [@sun_pymodes_2020], and platforms like The OpenSky Network [@opensky2014], have emerged, providing researchers with free access to real-time and historical aviation data.
 
 Most research initiatives and proofs of concept based on ADS-B data are limited to the use of historical data [@olive_traffic_2019]. However, the potential impact of such research depends on the possibility of adapting the algorithms to real-time data. The tangram framework aims to bridge this gap by providing a platform for real-time data collection and analysis, enabling researchers to conduct experiments and develop new algorithms with this real-time data efficiently.
 
@@ -53,7 +54,7 @@ The tangram framework consists of a suite of independent components that can be 
 
 The interaction between the frontend and backend is based on REST APIs and WebSocket connections. The REST API (the `tangram` application) is used for data retrieval and management. At the same time, the WebSocket connection (through the `channel` executable) allows for real-time communication between the frontend and backend.
 
-The backend is responsible for data collection, storage, and processing. All the components communicate through a Redis pub/sub system, which allows for efficient data exchange between components and real-time updates.
[…]
