### 2025-07-13 #8361 @anastassiavybornova
hi @xoolive holiday season indeed, delays to be expected, patience to be rewarded :) in the next couple of days, I will double-check the items on the list below, and aim to find at least 2 reviewers for your submission. as my colleague Sam explained above, it's the editor who selects & assigns reviewers, but you're very welcome to _suggest_ up to 5 reviewers from the [database listed above](https://reviewers.joss.theoj.org/lookup) or your own ([non-conflicted](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy)) extended network. Their github handles are most useful to receive **but **please don't use "@" to reference them** since it will prematurely ping them**.

**Software**
- [ ] ∃ software description
- [ ] ∃ OSI-approved license
- [ ] ∃ installation instructions
- [ ] ∃ documentation
- [ ] ∃ tests

- [ ] Paper compiles
- [ ] Paper length 250-1000 words

**Application**
- [ ] Substantial scholarly effort
