### 2025-07-13 #8361 @anastassiavybornova
@xoolive while i'm looking for reviewers, i'd like to ask you to do another round of shortening of the article text. i am currently getting a word count of ~1300, which is still quite above the typical limit of 250(!)-max 1000 words. you can consider using visual rather than text-based showcasing of features, and referring to the documentation for all omittable details.
