### 2025-07-13 #8361 @xoolive
Thanks @anastassiavybornova I'll try to reduce more after I'm back to office. Sorry I pinged people but I guess I can still suggest anyone among the same group of people?
