### 2025-07-15 #8361 @xoolive
Hi @anastassiavybornova

I had mentioned the following names:

- sfo and krumjan are contributors to another library we published on JOSS called traffic
- cathypacific8747 maintains a library to access similar sources of data, without the visualisation aspect addressed in tangram
- niclaswue has shown interest in the development efforts but has not participated in it
