### 2025-07-22 fbab666
reduce word count
