### 2025-07-22 #8361 @xoolive
Hi @anastassiavybornova I cut the manuscript by 350+ words now
