### 2025-07-22 #8361 @xoolive
I also see that you didn't tick the item for tests.

The main issue here is that the whole framework is based on real-time data. Some of the tools of the suite have their own test suites, like [jet1090](https://github.com/xoolive/rs1090/actions).

For the rest of the framework, the difficult part has been more to be able to produce a stable container image, which is automated [here](https://github.com/open-aviation/tangram/blob/main/.github/workflows/podman.yml) now.

Eventually, the final product works when the configuration is properly set (after the image is built), but we haven't found a proper way to address the problematic of unit tests per se as we usually perform them in offline processing software.
