### 2025-07-25 #8361 @fredshone
I can run through the checklist, but I'm not familiar with the subject/application, so will be more of a software review if that's ok?
