### 2025-07-29 #8662 @editorialbot
[REVIEW]: tangram, an open platform for modular, real-time air traffic management research
**Submitting author:** @xoolive (<a href="http://orcid.org/0000-0002-2335-5774">Xavier Olive</a>)
**Repository:** https://github.com/open-aviation/tangram
**Branch with paper.md** (empty if default branch):
**Version:** v0.1
