### 2025-07-30 open-aviation/tangram#36 @abc8747
joss review
Hi! I'll be leaving my comments for the review (https://github.com/openjournals/joss-reviews/issues/8662) here.

I'll be primarily working in the [`dev`](https://github.com/HKUST-OCTAD-LAB/AirTrafficSim/tree/dev) branch of the `AirTrafficSim` repo ([tracking issue](https://github.com/HKUST-OCTAD-LAB/AirTrafficSim/issues/65)). I have some internal refactoring for `AirTrafficSim` and will review tangram in depth tomorrow/Friday.
