### 2025-08-10 fd0a197
refactor!: transform monolithic application into a distributable plugin system (#36)
the primary motivation is to allow users to extend `tangram` with new data
sources (e.g., simulators) and visualisations without forking or modifying the
core codebase.

- backend: plugins are now pip-installable packages. the core backend discovers
  them at runtime using `importlib.metadata` entry points, decoupling from the
  main source tree.
- frontend: plugins are now standalone npm packages. each plugin builds its own
  assets and a `plugin.json` manifest. the backend generates a `manifest.json`
  based
[…]
