+++ packages/tangram_system/src/tangram_system/index.ts
@@ -1,0 +1,6 @@
+import type { TangramApi } from "@open-aviation/tangram/types";
+import SystemWidget from "./SystemWidget.vue";
+
+export function install(api: TangramApi) {
+  api.registerWidget("system", SystemWidget);
+}
