### 2025-08-12 e202fd1
refactor(core): compile `channel` service into python extension (#37)
embeds the Rust channel service directly into the python package using maturin
to simplify distribution and usage. this eliminates the need for `podman` and
`process-compose` for the channel service.

manylinux cibuildwheels will need to be configured in the future.
