+++ packages/tangram/rust/src/bin/stub_gen.rs
@@ -1,0 +1,11 @@
+// `cargo run --bin stub_gen --features python`
+use _channel::stub_info;
+use pyo3_stub_gen::Result;
+
+// finds the module name (`tangram._channel`) from pyproject.toml and generates
+// the pyi at `packages/tangram/src/tangram/_channel.pyi`
+fn main() -> Result<()> {
+    let stub = stub_info()?;
+    stub.generate()?;
+    Ok(())
+}

[…]
