### 2025-08-14 #8662 @fredshone
Firstly, congratulations on a lovely, well-made project @xoolive. I have two concerns with docs and tests...

Please keep in mind I am not a web developer, and as such, I might get this wrong. I am happy to discuss or debate any points. @anastassiavybornova it would be sensible to get someone with web dev experience to review this project.

**Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?**

- The quick start is excellent; however, I think you should add (i) an initial git clone instruction (🤷🏻) and (ii) instructions for correct shutdown, including removing containers/images.
- I think you should be clear from the start that jet1090 is a separately "published". There is sufficient contribution regardless.
- Similarly for channel, I think? Although I note channel is included in the [docs](https://mode-s.org/tangram/architecture/channel/), which is a little confusing. Seems like a particular risk of future incompatibility or confusion. I much prefer how the docs link to separate jet1090 docs...
- There is missing detail for [planes](https://mode-s.org/tangram/architecture/planes/) and [trajectory](https://mode-s.org/tangram/architecture/trajectory/). In all cases, I think you should add basic "dev" instructions, ie overview, install and testing. I think this is particularly important given the (excellent) modularity and extensibility of your project.
- I see both a py and rs version of planes,  which you refer to inconsistently in your paper figure and text.

**Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?**

- I don't see any tests? I am unclear to what extent it is possible to usefully test such a system, and I appreciate it will be time-consuming to make so many fake interfaces. Perhaps high-level integration tests would suffice? In any case, please provide and document.
