### 2025-08-21 ca465ef
updates for tangram review
