+++ paper/paper.bib
@@ -21,7 +21,8 @@
   booktitle = {Proceedings of the 13th International Symposium on {{Information}} Processing in Sensor Networks},
   author    = {Sch{\"a}fer, Matthias and Strohmeier, Martin and Lenders, Vincent and Martinovic, Ivan and Wilhelm, Matthias},
   year      = {2014},
-  pages     = {83--94}
+  pages     = {83--94},
+  doi       = {10.1109/ipsn.2014.6846743}
 }
 @inproceedings{olive_turbulence_2020,
   title     = {Detecting and {{Measuring Turbulence}} from {{Mode S Surveillance Downlink Data}}},
