+++ docs/architecture/planes.md
@@ -1,1 +1,23 @@
 # planes
+
+## Overview
+
+Planes is a specialized tool designed to monitor and track aircraft messages passing through the Redis pub/sub system. It maintains a comprehensive global view of the current air traffic situation by collecting and processing state vectors for all active aircraft (defined as those that have transmitted data within the past few minutes).
+
+## Key features
+
+- Continuous tracking of aircraft positions and metadata
+- Real-time state vector updates
+- Client-specific data filtering based on map view
+- Efficient network usage through targeted data distribution
+
+## How it works
+
+The system regularly publishes updated state vectors (once per second) on dedicated Redis channels for each client (individual web browser). Each client shares information about its current map view (bounding box), allowing `planes` to filter aircraft data and only send updates relevant to what each user can see on their screen. This approach significantly reduces network bandwidth requirements and improves overall system performance.
+
+## Implementations
+
+There are two available implementations of planes:
+
+- the Python implementation is used for development and quick implementations of improvements,
+- the stable Rust implementation is used as a _production_ version.
