### 2025-08-22 b719217
feat: prepare for new history plugin, remove podman (for now)
