+++ packages/tangram_history/src/tangram_history/__init__.py
@@ -90,8 +90,6 @@ async def get_track(icao24: str) -> list[dict[str, Any]]:
     return result
 
 
-# This function will be called by the main FastAPI application
-# Place it in __init__.py to register the plugin
 def register_plugin(app: FastAPI) -> None:
     """Register this plugin with the main FastAPI application."""
     app.include_router(router)
