+++ packages/tangram_jet1090/src/tangram_jet1090/_history_redis_ts.py
@@ -2,13 +2,16 @@
 # coding: utf8
 
 """
-Redis-based Aircraft History Tracking System
-===========================================
+!!! warning
+    deprecated, use `tangram_history` instead.
 
-This module implements an aircraft tracking system that uses Redis for state management
-and historical data storage. It subscribes to aircraft messages, processes them, and
-maintains both current state and historical positions of aircraft.
+Redis-based Aircraft History Tracking System with Time Series
+============================================================
 
+This module implements an aircraft tracking system that uses Redis Time Series for state
+management and historical data storage. It subscribes to aircraft messages, processes
+them, and maintains both current state and historical positions of aircraft using
+Redis's native time series capabilities.
 
 Redis Key Structure:
 -------------------
@@ -27,17 +30,24 @@ The system uses the following Redis key patterns:
    - Purpose: Tracks when we last wrote history for this aircraft to implement the
      aggregation interval
 
-3. Historical Position Data:
-   - Key: `aircraft:history:{icao24}:{timestamp}`
-   - Value: JSON encoded position history entry
-   - TTL: Configurable (default 10 minutes)
-   - Purpose: Stores individual position reports for historical tracks
+3. Time Series Data:
+   - Key patterns:
+     - `aircraft:ts:latitude:{icao24}`
+     - `aircraft:ts:longitude:{icao24}`
+     - `aircraft:ts:altitude:{icao24}`
+     - `aircraft:ts:track:{icao24}`
+   - Value: Time series of measurements
+   - Labels: Aircraft metadata (icao24, registration, typecode, callsign, type)
+   - Retention: Configurable (default 10 minutes)
+   - Purpose: Efficient storage and querying of time-based aircraft data
 
-4. Timeline Index:
-   - Key: `aircraft:timeline:{icao24}`
-   - Value: Sorted set mapping timestamps to scores
-   - TTL: Configurable (default 10 minutes)
-   - Purpose: Provides efficient time-based queries for aircraft tracks
+Benefits of Redis Time Series:
+----------------------------
+1. Memory Efficiency: Optimized storage for time series data
+2. Built-in Aggregation: Supports downsampling and aggregation functions
+3. Automatic Retention: Configurable data expiration
+4. Range Queries: Efficient retrieval of data by time range
+5. Labeled Data: Each time series can be tagged with metadata for easy filtering
 
 Usage:
 -----
@@ -56,7 +66,7 @@ You will want to apply `%autoawait` first to enable async support in ipython.
 
 - From within the tangram container:
 
-$ LOG_LEVEL=DEBUG uv run python -m tangram.history_redis
+$ LOG_LEVEL=DEBUG uv run python -m tangram.history_redis2
 
 - From Python REPL:
 
@@ -66,11 +76,11 @@ Use one of the following methods to start a REPL:
 - ipython with %autoawait
 - uv run ipython
 
-from tangram import history_redis
+from tangram import history_redis2
 
 # Create a client to interact with the aircraft data
 # if you REDIS_URL is set in environmental variable, you are good to go
-client = history_redis.StateClient()
+client = history_redis2.StateClient()
 
 # Get current aircraft
 count = await client.get_aircraft_count()
@@ -82,6 +92,10 @@ track = await client.get_aircraft_track("a0b1c2")
 # Get summary of active aircraft
 positions = await client.get_current_positions()
 
+# Get time series info and labels
+ts_info = await client.get_timeseries_info("a0b1c2")
+ts_labels = await client.get_timeseries_labels("a0b1c2")
+
 Useful Redis Commands:
 ---------------------
 Use redis-cli or iredis
[…]
