+++ docs/plugins/backend.md
@@ -138,8 +138,6 @@ async def get_example() -> ExampleResponse:
     return ExampleResponse(data="This is an example plugin response")
 
 
-# This function will be called by the main FastAPI application
-# Place it in __init__.py to register the plugin
 def register_plugin(app: FastAPI) -> None:
     """Register this plugin with the main FastAPI application."""
     app.include_router(router)
