### 2025-08-23 a07ba58
refactor: move `planes` binary into `packages/tangram-jet1090` as cdylib
the core is now responsible for discovering downstream managed services and
spawning a new thread for them
