+++ packages/tangram_jet1090/rust/src/bin/stub_gen.rs
@@ -1,0 +1,8 @@
+use _planes::stub_info;
+use pyo3_stub_gen::Result;
+
+fn main() -> Result<()> {
+    let stub = stub_info()?;
+    stub.generate()?;
+    Ok(())
+}
