+++ packages/tangram/rust/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "channel_python"
+name = "channel"
 version = "0.2.8"
 edition = "2021"
 
@@ -28,7 +28,7 @@ pyo3-stub-gen = { version = "0.12.2", optional = true }
 
 
 [[bin]]
-name = "stub_gen"
+name = "stub_gen_channel"
 path = "src/bin/stub_gen.rs"
 
 [features]
