### 2025-08-24 24b799a
chore(ci): bdist wheels, dummy podman containerfile and compose
