### 2025-08-24 170ebba
feat: basic tests for core package
