+++ scripts/tangram.toml
@@ -1,8 +1,0 @@
-[core]
-plugins = [
-    "tangram_example",
-    "tangram_jet1090",
-    "tangram_history",
-    "tangram_system",
-    "tangram_weather"
-]
