### 2025-08-25 9024b4d
ci: make podman more robust, fix vite bugs
- build eccodes from scratch, regardless of the architecture (for now)
- use `uv sync --all-packages` instead in container
- for tests and wheel building, we continue to use maturin + hatch instead of
  `uv sync`.
- the new frontend now uses exports conditions with a custom "source" condition
  for simplicity.

in the future, we should probably standardise whether we want to do `uv sync`
(which is more convenient) or `uv build`+`maturin build` (more "correct",
seemingly lower binary size and greater performance). we should also consider
the environmen
[…]
