+++ pyproject.toml
@@ -2,41 +2,26 @@
 name = "tangram_workspace"
 version = "0.2.0"
 requires-python = ">=3.10"
-# dependencies specified here are for ci / local development only
-dependencies = ["tangram"]
+dependencies = []
 classifiers = [
     "Private :: Do Not Upload"
 ]
 
-[project.optional-dependencies]
-example = ["tangram-example"]
-history = ["tangram-history"]
-jet1090 = ["tangram-jet1090"]
-system = ["tangram-system"]
-weather = ["tangram-weather"]
-
 [dependency-groups]
 dev = [
     "ipykernel>=6.29.5",
-    "maturin>=1.9.3",
+    { include-group = "build" },
     { include-group = "test" },
     { include-group = "typing" },
     { include-group = "lint" },
 ]
+build = ["maturin[patchelf]>=1.9.3"]
 test = ["pytest>=8.4.1", "anyio>=4.8.0"]
 lint = ["ruff>=0.12.7"]
 typing = ["mypy>=1.17.1"]
 
 [tool.uv.workspace]
 members = ["packages/*"]
-
-[tool.uv.sources]
-tangram = { workspace = true }
-tangram-example = { workspace = true }
-tangram-history = { workspace = true }
-tangram-jet1090 = { workspace = true }
-tangram-system = { workspace = true }
-tangram-weather = { workspace = true }
 
 [tool.mypy]
 python_version = "3.10"
