### 2025-08-25 ed65357
ci: fix vite typescript bug, add more platforms
- spawn tangram server directory within tests
- use vendored openssl on specific platforms to resolve maturin build failures
- build eccodes from scratch on aarch64
