### 2025-08-29 a8265c7
docs: quickstart - ci: tiny fix for maturin wheels
