+++ mkdocs.yml
@@ -36,8 +36,6 @@ theme:
     - navigation.expand
     - navigation.footer
     - navigation.indexes
-    - navigation.tabs
-    - navigation.tabs.sticky
     - search.highlight
     - search.suggest
   palette:
