+++ .github/workflows/ci.yml
@@ -100,7 +100,7 @@ jobs:
           - os: linux
             manylinux: manylinux_2_28
             target: aarch64
-            features: --features openssl-vendored
+            features: --features openssl-vendored python
           - os: macos
             target: x86_64
           - os: macos
@@ -115,34 +115,34 @@ jobs:
           - os: linux
             manylinux: manylinux_2_28
             target: aarch64
-            features: --features openssl-vendored
+            features: --features openssl-vendored python
           - os: linux
             manylinux: manylinux_2_28
             target: armv7
             interpreter: 3.10 3.11 3.12 3.13 3.14
-            features: --features openssl-vendored
+            features: --features openssl-vendored python
           - os: linux
             manylinux: manylinux_2_28
             target: ppc64le
             interpreter: 3.10 3.11 3.12 3.13 3.14
-            features: --features openssl-vendored
+            features: --features openssl-vendored python
           - os: linux
             manylinux: manylinux_2_28
             target: s390x
             interpreter: 3.10 3.11 3.12 3.13 3.14
-            features: --features openssl-vendored
+            features: --features openssl-vendored python
           - os: linux
             manylinux: musllinux_1_1
             target: x86_64
-            features: --features openssl-vendored
+            features: --features openssl-vendored python
           - os: linux
             manylinux: musllinux_1_1
             target: aarch64
-            features: --features openssl-vendored
+            features: --features openssl-vendored python
           - os: linux
             manylinux: musllinux_1_1
             target: armv7
-            features: --features openssl-vendored
+            features: --features openssl-vendored python
 
     runs-on: ${{ (matrix.os == 'linux' && 'ubuntu') || matrix.os }}-latest
     steps:
