### 2025-08-31 9e93008
docs: align with new backend plugins system, rename `web/` to `web_legacy`
deprecated documentation are put in comments (for now) and named with prefix `_`.
the split between jet1090/history still needs some work.

ci: quick fix for maturin/rust features
