+++ docs/architecture/channel.md
@@ -117,32 +117,23 @@ for message in p.listen():
 
 ## Running the Channel Service
 
-The Channel service is containerized for easy deployment:
+The Channel service is an integrated part of the core `tangram` application. It is automatically started as a background service when you run the `tangram serve` command. You do not need to run it separately.
 
-```bash
-# Pull the Channel image
-podman pull ghcr.io/emctoo/channel:latest
+Its behavior is configured in the `[channel]` section of your `tangram.toml` file.
 
-# Run the Channel service
-podman run -it --rm --name channel --network host \
-  ghcr.io/emctoo/channel:latest \
-  channel --redis-url redis://localhost:6379 --jwt-secret your-secret-key
+```toml title="tangram.toml"
+[channel]
+host = "127.0.0.1"
+port = 2347
+jwt_secret = "a-better-secret-than-this"
 ```
-
-### Configuration Options
-
-- `--host`: Host address to bind to (default: 127.0.0.1)
-- `--port`: Port to listen on (default: 2025)
-- `--redis-url`: Redis connection URL
-- `--jwt-secret`: Secret key for JWT authentication
-- `--static-path`: Path to static assets directory (default: assets)
 
 ## Token Authentication
 
-Channels require JWT tokens for authentication. Tokens can be requested from the `/token` endpoint:
+Channels require JWT tokens for authentication. Tokens can be requested from the `/token` endpoint, which is exposed on the port defined in your configuration.
 
 ```bash
-curl -X POST http://localhost:2025/token \
+curl -X POST http://localhost:2347/token \
   -H "Content-Type: application/json" \
   -d '{"channel": "system", "id": "client1"}'
 ```
@@ -157,17 +148,6 @@ The response includes a JWT token that can be used for channel authentication:
 }
 ```
 
-## Deployment, Testing and Debugging
-
-During development, the channel service operates within the container through process-manager (defined in `../container/process-compose.yaml`). To enable browser communication with this service, a WebSocket proxy is configured in `../web/vite.config.js`. This proxy prioritizes the endpoint specified in the `CHANNEL_SERVICE` environment variable, defaulting to `localhost:2025` within the container if no value is provided. For more flexible deployment options, you can run the service on a separate node and specify its location in the `.env` file.
-
-The Channel service includes an admin interface for monitoring connections and messages:
-
-- **Admin Interface**: Available at `http://localhost:2025/admin.html`
-- **Client Test Page**: Available at `http://localhost:2025?name=testuser`
-
-(FIXME: `rust-embed` for assets embedding)
-
 ### Redis Commands for Debugging
 
 ```bash
@@ -180,24 +160,3 @@ redis-cli psubscribe "to:*"
 # Publish a test message to clients
 redis-cli publish "to:system:test" '{"type":"message","message":"Test from Redis"}'
 ```
-
-## Progressive Integration in Tangram
-
-The Channel component enables the progressive development approach in Tangram:
-
-1. **Basic API Integration**: Start with conventional REST API calls
-2. **Enhanced Processing**: Add WebSocket capabilities for specific real-time features
-3. **Real-time Streaming**: Fully leverage the WebSocket infrastructure for comprehensive real-time visualization
-
-By separating the communication layer from both the frontend and backend, Channel allows teams to adopt real-time capabilities at their own pace while maintaining backward compatibility with existing systems.
-
-### Channel (Streaming Service)
-
-The Channel component is a high-performance Rust implementation that handles real-time data streaming:
-
-- WebSocket server for frontend connections
-- Integration with Redis for pub/sub messaging
-- Efficient data serialization and transmission
-- Connection management and error handling
-
-This component enables the real-time nature of Tangram, allowing for immediate visualization of incoming aviation data.
