+++ docs/plugins/jet1090.md
@@ -1,0 +1,36 @@
+# Jet1090 Plugin
+
+The `tangram_jet1090` plugin is the primary tool for integrating live Mode S and ADS-B data into the `tangram` framework. It processes raw aviation surveillance data from a `jet1090` instance and makes it available for real-time visualization and historical analysis.
+
+## Overview
+
+This plugin consists of:
+
+- **A background service** (`planes`) to maintain a real-time state of all visible aircraft.
+- **A REST API endpoint** (`/data/{icao24}`) to fetch raw message history for a specific aircraft.
+- **A frontend widget** to display a placeholder in the UI.
+
+## `planes` Service
+
+The `planes` service is the core of the real-time functionality. It is a Rust-based component, wrapped with PyO3 for integration into the Python ecosystem.
+
+- **Continuous Tracking**: It subscribes to the `jet1090` Redis channel to receive decoded aircraft messages.
+- **State Vector Maintenance**: It maintains a comprehensive in-memory view of the current air traffic situation by collecting and processing state vectors for all active aircraft. A state vector is a data structure containing the latest known information for each aircraft parameter (position, altitude, speed, etc.), as this data arrives in separate messages.
+- **Client-Specific Filtering**: The service listens for bounding box updates from each connected frontend client. It filters the aircraft data for each client, sending only the aircraft visible within their map view.
+- **Data Publishing**: Once per second, it publishes the filtered state vectors to a dedicated Redis channel for each client (e.g., `to:streaming-<client_id>:new-data`), which are then relayed to the browser via the WebSocket `channel` service.
+
+## Configuration
+
+To use this plugin, you must have a running `jet1090` instance publishing data to Redis. You can also configure the plugin in your `tangram.toml`:
+
+```toml
+[core]
+plugins = ["tangram_jet1090"]
+
+[plugins.tangram_jet1090]
+# The Redis channel that jet1090 is publishing to.
+jet1090_channel = "jet1090"
+# How long (in seconds) to keep an aircraft in the state vector table
+# after its last message.
+history_expire = 20
+```
