+++ docs/plugins/history.md
@@ -1,0 +1,3 @@
+The `tangram_history` plugin provides an API for querying historical aircraft data, such as full flight trajectories.
+
+<!-- TODO: we should move jet1090 components here -->

[…]
