### 2025-09-01 22dd7b0
docs: add api reference
adds intersphinx support and enables cross linking in the future

ci: skip s390x and remove duplicate aarch64
