+++ packages/tangram_jet1090/src/tangram_jet1090/_history_redis.py
@@ -114,9 +114,9 @@ Monitoring:
 import asyncio
 import logging
 import os
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 from datetime import datetime, timezone
-from typing import TYPE_CHECKING, Any, Dict, ForwardRef, List, Optional
+from typing import Any, Dict, List, Optional
 
 import msgspec
 import rs1090
