+++ docs/plugins/history.md
@@ -1,3 +1,6 @@
-The `tangram_history` plugin provides an API for querying historical aircraft data, such as full flight trajectories.
+# History Plugin
 
-<!-- TODO: we should move jet1090 components here -->
+!!! note
+    This plugin is under active development, see the [jet1090 plugin](./jet1090.md) for more information.
+
+The `tangram_history` plugin provides an API for querying historical aircraft data, such as full flight trajectories.
