+++ pyproject.toml
@@ -19,6 +19,10 @@ build = ["maturin[patchelf]>=1.9.3"]
 test = ["pytest>=8.4.1", "anyio>=4.8.0"]
 lint = ["ruff>=0.12.7"]
 typing = ["mypy>=1.17.1"]
+docs = [
+    "mkdocs-material[imaging]>=9.6.18",
+    "mkdocstrings-python>=1.18.2",
+]
 
 [tool.uv.workspace]
 members = ["packages/*"]
@@ -41,6 +45,7 @@ warn_unused_ignores = true
 [tool.ruff]
 line-length = 88
 target-version = "py310"
+extend-exclude = ["*.pyi"]  # don't format files generated by pyo3_stub_gen
 
 [tool.ruff.lint]
 select = [
