### 2025-09-01 #8662 @fredshone
@xoolive apologies for the slow response, I've been on holiday.

re the quickstart, I see the instruction `podman container stop redis`. But not for other containers? Maybe just add `podman system prune` at the end? Otherwise quickstart looks great.

Understood and accepted re the location of code and docs for tangram, rs1090, and channels.

re rs1090 tests, I have since seen that rs1090 has some coverage for the python bindings. There is no documentation about running these tests, but i see them in the CI, so I think this **complies** with JOSS checklist. For the benefit of future users and maintainers, I **suggest** you also add some info on these tests to the docs and consider some rust-level coverage.

re tangram and channels, I see no coverage other than a CI check for a successful build? I think the easiest route to comply with the review checklist is to document that the best way to test is to build, run and verify manually (_Are there automated tests or manual steps described so that the functionality of the software can be verified?_). Otherwise, there do look like opportunities to add in module-level unit testing, but I appreciate that this doesn't well cover the overall application, so not a requirement. Given the amount of work you have clearly done, faking some interfaces to get more thorough and rapid assurance seems worthwhile, but I leave it to personal choice.

To clarify, only two required actions:
- [ ]  podman prune of similar at end of quickstart
- [ ] testing tangram/channels - at least some docs (probably just a link to the quickstart)
