### 2025-09-01 f8f647b
refactor: unify frontend path, routers and services in a single class
- fix: frontend not being located when package is installed as editable
- fix: hardcoded jet1090 channel config
- feat: `tangram list-plugins [--config [/path/to/tangram.toml]|--all]
example with only example and jet1090 enabled:
┏━━━━━━━━━━━━━━━━━┳━━━━━━━━━━━┳━━━━━━━━━━┳━━━━━━━━━━┳━━━━━━━━━━━━┓
┃ plugin          ┃ status    ┃ frontend ┃ routers  ┃ services   ┃
┡━━━━━━━━━━━━━━━━━╇━━━━━━━━━━━╇━━━━━━━━━━╇━━━━━━━━━━╇━━━━━━━━━━━━┩
│ tangram_weather │ available │ ?        │ ?        │ ?          │
│ tangram_example │ enabled   │ 
[…]
