+++ packages/tangram/src/tangram/__init__.py
@@ -1,0 +1,4 @@
+from .config import Config
+from .plugin import Plugin
+
+__all__ = ["Config", "Plugin"]
