+++ packages/tangram/src/tangram/__main__.py
@@ -1,108 +1,39 @@
 import asyncio
-import importlib.metadata
-import importlib.resources
 import logging
 from importlib.abc import Traversable
 from pathlib import Path
-from typing import Annotated, AsyncGenerator
+from typing import Annotated
 
 import typer
 import uvicorn
-from fastapi.responses import JSONResponse
-from fastapi.staticfiles import StaticFiles
+from rich.console import Console
+from rich.table import Table
 
-from .backend import create_app
-from .config import TangramConfig
+from .backend import create_app, load_enabled_plugins, resolve_frontend, run_services
+from .config import Config
+from .plugin import DistName, Plugin, load_plugin, scan_plugins
 
 app = typer.Typer(no_args_is_help=True)
 logger = logging.getLogger(__name__)
-
-
-async def run_channel_service(config: TangramConfig) -> None:
-    from ._channel import ChannelConfig, init_logging, run
-
-    init_logging("debug")
-
-    rust_config = ChannelConfig(
-        host=config.channel.host,
-        port=config.channel.port,
-        redis_url=config.core.redis_url,
-        jwt_secret=config.channel.jwt_secret,
-        jwt_expiration_secs=config.channel.jwt_expiration_secs,
-    )
-    await run(rust_config)
-
-
-async def run_services(
-    config: TangramConfig,
-) -> AsyncGenerator[asyncio.Task[None], None]:
-    yield asyncio.create_task(run_channel_service(config))
-    for entry_point in importlib.metadata.entry_points(group="tangram.services"):
-        try:
-            service_run_func = entry_point.load()
-            yield asyncio.create_task(service_run_func(config))
-            logger.info(f"started service: {entry_point.name}")
-        except Exception as e:
-            logger.error(f"failed to load service {entry_point.name}: {e}")
+stderr = Console(stderr=True)
 
 
 async def run_server(
-    config: TangramConfig, host: str, port: int, static_dir: Traversable
+    config: Config, loaded_plugins: list[tuple[DistName, Plugin]]
 ) -> None:
-    app_instance = create_app()
-    app_instance.state.config = config
-
-    frontend_plugins = []
-    for plugin_name in config.core.plugins:
-        try:
-            plugin_dist = importlib.resources.files(plugin_name) / "dist-frontend"
-        except ModuleNotFoundError:
-            logger.warning(
-                f"expected plugin `{plugin_name}` to be installed, but it was not; "
-                "skipping."
-            )
-            continue
-        if not plugin_dist.is_dir():
-            # FIXME(abr): this currently happens on `uv sync --all-packages`, which
-            # installs the project as *editable*, but:
-            # 1. vite is configured to write to packages/{plugin}/dist-frontend
-            # 2. hatchling and maturin, on wheel building, properly adds it to the wheel
-            # 3. `importlib.resources` return the path to the root of the bdist wheel,
-            #    which works for `pip`-installed packages,
-            #    but for editable, `importlib` returns packages/{plugin}/src/{plugin}/
-            #    whose dist-frontend is missing.
-            logger.warning(
-                f"expected plugin `{plugin_name}` to have a frontend at {plugin_dist}, "
-                "but it was not found; skipping."
-            )
-            continue
-        app_instance.mount(
-            f"/plugins/{plugin_name}",
-            StaticFiles(directory=str(plugin_dist)),
-            name=plugin_name,
-        )
-        frontend_plugins.append(plugin_name)
-
-    @app_instance.get("/manifest.json")
-    async def get_manifest() -> JSONResponse:
-        return JSONResponse(content={"plugins": frontend_plugins})
-
-    app_instance.mount(
[…]
