+++ packages/tangram_example/readme.md
@@ -4,9 +4,8 @@ This is an example plugin for the tangram framework.
 
 It serves as a minimal template for creating new backend plugins for API services.
 
-The procedure is as simple as:
+A plugin is a standard Python package that uses entry points in its `pyproject.toml` to make itself discoverable by `tangram`. See the `[project.entry-points."tangram.plugins"]` section for an example.
 
-- Create a new directory in the `src/tangram/plugins/` directory.
-- Add the plugin to the `src/tangram/plugins/__init__.py` file: it will need a `router` instance and a `register_plugin` function. The `register_plugin` will be called when the module is discovered at runtime.
+The plugin's `__init__.py` should define a `plugin = tangram.Plugin()` instance and use its decorators (`@plugin.register_router`, `@plugin.register_service`) to register components.
 
-You can explore more plugins in the `src/tangram/plugins/` directory to see how they are structured. A more complete documentation is available at <https://mode-s.org/tangram/plugins/>
+You can explore more plugins in the `packages/*` directory to see how they are structured. A more complete documentation is available at <https://mode-s.org/tangram/plugins/>

[…]
