### 2025-09-02 276b30e
refactor: share single redis client across plugins - refactor: rename `python` to `pyo3`
