+++ packages/tangram/src/tangram/__init__.py
@@ -1,4 +1,5 @@
+from .backend import BackendState, InjectBackendState
 from .config import Config
 from .plugin import Plugin
 
-__all__ = ["Config", "Plugin"]
+__all__ = ["BackendState", "Config", "InjectBackendState", "Plugin"]
