### 2025-09-02 750d72b
refactor: move legacy jet1090 planes/trajectory code into new directory
