+++ docs/architecture/index.md
@@ -26,15 +26,14 @@ This design allows you to:
 
 ![tangram architecture](../screenshot/tangram_diagram.png)
 
-| **Component**         | **Provided By**          | **Description**                                            |
-| --------------------- | ------------------------ | ---------------------------------------------------------- |
-| `tangram` (Core)      | `tangram` package        | REST API server, CLI, and frontend shell.                  |
-| `channel`             | (Bundled with `tangram`) | WebSocket bridge between the frontend and Redis pub/sub.   |
-| `jet1090` integration | `tangram_jet1090` plugin | Decodes Mode S/ADS-B messages and provides data streams.   |
-| State Vectors         | `tangram_jet1090` plugin | Maintains a real-time state table of aircraft.             |
-| Trajectory History    | `tangram_history` plugin | Stores and retrieves historical aircraft data.             |
-| System Info           | `tangram_system` plugin  | Provides backend server metrics like CPU and memory usage. |
-| Weather Layers        | `tangram_weather` plugin |                                                            |
+| **Component**            | **Provided By**          | **Description**                                                |
+| ------------------------ | ------------------------ | -------------------------------------------------------------- |
+| `tangram` (Core)         | `tangram` package        | REST API server, CLI, and frontend shell.                      |
+| `channel`                | (Bundled with `tangram`) | WebSocket bridge between the frontend and Redis pub/sub.       |
+| `jet1090` integration    | `tangram_jet1090` plugin | Decodes Mode S/ADS-B messages and provides data streams.       |
+| State Vectors & History  | `tangram_jet1090` plugin | Maintains real-time state and stores historical aircraft data. |
+| System Info              | `tangram_system` plugin  | Provides backend server metrics like CPU and memory usage.     |
+| Weather Layers           | `tangram_weather` plugin | Provides API endpoints for meteorological data.                |
 
 ## Backend Plugin System
 
@@ -42,74 +41,8 @@ The backend discovers plugins using Python's standard **entry point** mechanism.
 
 For a detailed guide on creating your own backend extensions, see the [Backend Plugin Guide](../plugins/backend.md).
 
-
-<!-- ### jet1090
-
-**Documentation**: <https://mode-s.org/jet1090>
-
-`jet1090` is a Rust-based ADS-B decoder that can be used to decode ADS-B messages from a variety of sources, including software-defined radio devices and network streams such as those provided by commercial ADS-B receivers. It is designed to be fast and efficient, making it suitable for real-time applications.
-
-### planes
-
-**Documentation**: [planes](planes.md)
-
-`planes` is a Python-based component that maintains a state vector table of aircraft. It is responsible for tracking the position and other parameters of aircraft in real-time. The component uses the data provided by `jet1090` to update the state vector table and provide real-time information about the aircraft.
-
-A state vector table is a data structure that contains the latest information about all the parameters of the aircraft. This is necessary because all ADS-B messages do not provide all the parameters of the aircraft at once. In particular the position, the speed, the track angle and the identification all come in different messages. Having the most recent information along all the possible features is usually enough to display the aircraft on a map.
-
-### trajectory
-
-**Documentation**: [trajectory](trajectory.md)
-
-`trajectory` is a Python-based component that provides a history of data for a given aircraft. It is responsible for storing and retrieving historical data about the aircraft's position and other parameters. The component uses the data stored by the Redis system and reformat it to be used by the frontend in a more standard JSON-like format. This is useful to be able to display the trajectory of an aircraft on a map, as well as to provide historical data for plotting, e.g. altitude, speed, vertical rate, etc.
-
-### tangram REST API
-
-**Documentation**: available when you run the service at <http://localhost:2345/tangram/docs>
-
[…]

[…]
