### 2025-09-02 58c7a43
fix: update default backend port to 2346, bring back legacy frontend .env.example - remove compose.yml (we can dynamically generate it from tangram.toml) anyways
