+++ packages/tangram/src/tangram/config.py
@@ -9,7 +9,7 @@ from typing import Any
 @dataclass
 class ServerConfig:
     host: str = "127.0.0.1"
-    port: int = 8000
+    port: int = 2346
 
 
 @dataclass
