+++ tangram.example.toml
@@ -9,7 +9,7 @@ plugins = [
 
 [server]
 host = "0.0.0.0"
-port = 8000
+port = 2346
 
 [channel]
 host = "0.0.0.0"
