### 2025-09-03 6d571cf
fix(jet1090): crash due to repeated creations of multiplexed redis connection - fix(jet1090): backend events for `system` and `admin` were dropped because   their listener tasks were never spawned at startup - fix(frontend): speed not showing on chart, trajectory not showing on map - fix(frontend): clear old chart data immediately to prevent showing stale version - perf(jet1090): use ts mrange to fetch all available time series for an aircraft   in a single network request, handling missing keys naturally - chore: instruct uv to recompile on rust code changes
