### 2025-09-03 9955e8b
feat: forward rust tracing data to `Layer` defined in and passed from python
this is less performant than simply logging to stdout on the rust side, but the
advantage is that we get a centralised place to configure logging across plugins
