+++ packages/tangram/src/tangram/__main__.py
@@ -6,6 +6,7 @@ from typing import Annotated
 
 import typer
 from rich.console import Console
+from rich.logging import RichHandler
 from rich.table import Table
 
 from .backend import resolve_frontend, start_tasks
@@ -22,6 +23,13 @@ def serve(
     config: Annotated[Path, typer.Option(help="Path to the tangram.toml config file.")],
 ) -> None:
     """Serves the core tangram frontend and backend services."""
+    logging.basicConfig(
+        level="NOTSET",
+        format="%(message)s",
+        datefmt="[%X]",
+        handlers=[RichHandler()],
+    )
+
     if not config.is_file():
         logger.error(f"config file not found: {config}")
         raise typer.Exit()
