+++ packages/tangram/src/tangram/__init__.py
@@ -1,5 +1,5 @@
-from .backend import BackendState, InjectBackendState
+from .backend import BackendState, InjectBackendState, TracingLayer
 from .config import Config
 from .plugin import Plugin
 
-__all__ = ["BackendState", "Config", "InjectBackendState", "Plugin"]
+__all__ = ["BackendState", "Config", "InjectBackendState", "Plugin", "TracingLayer"]
