### 2025-09-04 be58daa
perf: use env filter to avoid spamming python with `TRACE` - disable python tracing by default (because of perf) and make it configurable - we might want to think about how to write logs to a file on both the python   and rust side
