### 2025-09-04 bf3276e
refactor: expose more configuration options (jet1090 stream interval...)
