### 2025-09-04 15930b0
refactor: unify logging formatting - fix: avoid hanging entire restapi when weather is downloading - fix: handle ctrl+c gracefully
