### 2025-09-05 ce0dbd6
docs: improve cross references, use mermaid - chore: add authors to each package from blame
