+++ docs/api_reference.md
@@ -1,7 +1,0 @@
-::: tangram
-    options:
-        show_if_no_docstring: true
-        show_submodules: true
-        filters:
-            - "!^_"
-            - "tangram.__main__"
