### 2025-09-05 285a551
support out-of-tree plugins (backend) (#37)
Co-authored-by: Xavier Olive <git@xoolive.org>
