+++ crates/planes/src/main.rs
@@ -1,542 +1,0 @@
-mod aircraftdb;
-
-// TODO include this part in the rs1090 crate rather than jet1090
-use crate::aircraftdb::{aircraft, Aircraft};
-use anyhow::{Context, Result};
-use clap::Parser;
-use futures::StreamExt;
-use redis::{AsyncCommands, RedisResult};
-use rs1090::data::patterns::aircraft_information;
-use serde::{Deserialize, Serialize};
-use std::collections::BTreeMap;
-use std::iter::Iterator;
-use std::{
-    collections::{HashMap, HashSet},
-    sync::Arc,
-    time::{Duration, SystemTime, UNIX_EPOCH},
-};
-use tokio::{sync::Mutex, time};
-use tracing::{debug, error, info};
-use tracing_subscriber::EnvFilter;
-
-// Command-line arguments
-#[derive(Parser, Debug)]
-#[clap(author, version, about = "Aircraft data streaming service")]
-struct Args {
-    /// Redis URL
-    #[clap(long, env = "REDIS_URL", default_value = "redis://redis:6379")]
-    redis_url: String,
-
-    /// Redis channel for Jet1090 messages
-    #[clap(long, env = "JET1090_CHANNEL", default_value = "jet1090")]
-    jet1090_channel: String,
-
-    /// Expire aircraft after (in seconds)
-    #[clap(long, env = "EXPIRE_AIRCRAFT", default_value = None)]
-    expire: Option<u16>,
-}
-
-// Struct for bounding box data
-#[derive(Debug, Clone, Serialize, Deserialize)]
-struct BoundingBox {
-    north_east_lat: f64,
-    north_east_lng: f64,
-    south_west_lat: f64,
-    south_west_lng: f64,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-struct StateVector {
-    icao24: String,
-    registration: Option<String>,
-    typecode: Option<String>,
-    lastseen: f64,
-    firstseen: f64,
-    callsign: Option<String>,
-    latitude: Option<f64>,
-    longitude: Option<f64>,
-    altitude: Option<f64>,
-    track: Option<f64>,
-}
-
-// Message from Redis
-#[derive(Debug, Clone, Serialize, Deserialize)]
-struct ClientMessage {
-    #[serde(rename = "connectionId")]
-    connection_id: String,
-    #[serde(flatten)]
-    data: HashMap<String, serde_json::Value>,
-}
-
-// Bounding box message
-#[derive(Debug, Clone, Serialize, Deserialize)]
-struct BoundingBoxMessage {
-    #[serde(rename = "connectionId")]
-    connection_id: String,
-    #[serde(rename = "northEastLat")]
-    north_east_lat: f64,
-    #[serde(rename = "northEastLng")]
-    north_east_lng: f64,
-    #[serde(rename = "southWestLat")]
-    south_west_lat: f64,
-    #[serde(rename = "southWestLng")]
-    south_west_lng: f64,
-}
-
-/// Jet1090 message format
-#[derive(Debug, Clone, Serialize, Deserialize)]
-struct Jet1090Message {
-    icao24: String,
-    df: String,
-    bds: Option<String>,
-    timestamp: f64,
-    callsign: Option<String>,
-    altitude: Option<f64>,
-    latitude: Option<f64>,
-    longitude: Option<f64>,
-    track: Option<f64>,
-    #[serde(flatten)]
-    extra: HashMap<String, serde_json::Value>,
-}
-
-/// Response data to clients
-#[derive(Debug, Clone, Serialize, Deserialize)]
-struct ResponseData {
-    count: usize,
-    aircraft: Vec<StateVector>,
-}
-
-/// State vectors collection to track aircraft
[…]
