+++ packages/tangram_weather/src/tangram_weather/index.ts
@@ -1,0 +1,6 @@
+import type { TangramApi } from "@open-aviation/tangram/types";
+import WeatherWidget from "./WeatherWidget.vue";
+
+export function install(api: TangramApi) {
+  api.registerWidget("weather", WeatherWidget);
+}
