+++ packages/tangram_jet1090/rust/src/lib.rs
@@ -1,0 +1,187 @@
+pub mod aircraftdb;
+pub mod bbox;
+pub mod state;
+pub mod stream;
+
+use anyhow::{Context, Result};
+#[cfg(feature = "pyo3")]
+use pyo3::{
+    exceptions::{PyOSError, PyRuntimeError},
+    prelude::*,
+};
+#[cfg(feature = "pyo3")]
+use pyo3_stub_gen::derive::*;
+use std::sync::Arc;
+use tokio::sync::Mutex;
+use tracing::{error, info};
+
+use crate::bbox::BoundingBoxState;
+use crate::state::StateVectors;
+use crate::stream::{start_jet1090_subscriber, start_redis_subscriber, stream_statevectors};
+#[cfg(feature = "pyo3")]
+use pyo3_python_tracing_subscriber::PythonCallbackLayerBridge;
+#[cfg(feature = "pyo3")]
+use tracing_subscriber::{fmt, EnvFilter, prelude::*};
+
+#[cfg(feature = "pyo3")]
+#[gen_stub_pyfunction]
+#[pyfunction]
+fn init_tracing_python(py_layer: Bound<'_, PyAny>, filter_str: String) -> PyResult<()> {
+    tracing_subscriber::registry()
+        .with(EnvFilter::new(filter_str))
+        .with(PythonCallbackLayerBridge::new(py_layer))
+        .try_init()
+        .map_err(|e| PyOSError::new_err(e.to_string()))
+}
+
+#[cfg(feature = "pyo3")]
+#[gen_stub_pyfunction]
+#[pyfunction]
+fn init_tracing_stderr(filter_str: String) -> PyResult<()> {
+    tracing_subscriber::registry()
+        .with(EnvFilter::new(filter_str))
+        .with(fmt::layer().with_writer(std::io::stderr))
+        .try_init()
+        .map_err(|e| PyOSError::new_err(e.to_string()))
+}
+
+#[cfg_attr(feature = "pyo3", gen_stub_pyclass)]
+#[cfg_attr(feature = "pyo3", pyclass(get_all, set_all))]
+#[derive(Debug, Clone)]
+pub struct PlanesConfig {
+    pub redis_url: String,
+    pub jet1090_channel: String,
+    pub history_expire: u16,
+    pub stream_interval_secs: f64,
+    pub aircraft_db_url: String,
+    pub aircraft_db_cache_path: Option<String>,
+}
+
+#[cfg(feature = "pyo3")]
+#[gen_stub_pymethods]
+#[pymethods]
+impl PlanesConfig {
+    #[new]
+    fn new(
+        redis_url: String,
+        jet1090_channel: String,
+        history_expire: u16,
+        stream_interval_secs: f64,
+        aircraft_db_url: String,
+        aircraft_db_cache_path: Option<String>,
+    ) -> Self {
+        Self {
+            redis_url,
+            jet1090_channel,
+            history_expire,
+            stream_interval_secs,
+            aircraft_db_url,
+            aircraft_db_cache_path,
+        }
+    }
+}
+
+async fn _run_service(config: PlanesConfig) -> Result<()> {
+    let bbox_state = Arc::new(Mutex::new(BoundingBoxState::new()));
+    let bbox_subscriber_state = Arc::clone(&bbox_state);
+
+    let redis_url_clone1 = config.redis_url.clone();
+    let bbox_subscriber_handle = tokio::spawn(async move {
+        match start_redis_subscriber(redis_url_clone1, bbox_subscriber_state).await {
+            Ok(_) => info!("BoundingBox subscriber stopped normally"),
+            Err(e) => error!("BoundingBox subscriber error: {}", e),
+        }
+    });
+
+    let client = redis::Client::open(config.redis_url.clone())
[…]
