### 2025-09-05 23a3eee
support out-of-tree plugins (backend) (#37)
Co-authored-by: Xavier Olive <git@xoolive.org>
