+++ src/tangram/history.py
@@ -1,187 +1,0 @@
-import asyncio
-import json
-import logging
-from dataclasses import dataclass
-from typing import Any, List
-
-from redis.commands.timeseries import TimeSeries
-
-from tangram.common import database, redis
-
-logging.basicConfig(
-    level=logging.INFO, format="%(asctime)s - %(levelname)s - %(message)s"
-)
-log = logging.getLogger(__name__)
-
-
-@dataclass
-class State:
-    icao24: str | None = None
-
-
-class Subscriber(redis.Subscriber[State]):
-    def __init__(
-        self,
-        name: str,
-        redis_url: str,
-        channels: List[str],
-        history_db: database.StateVectorDB,
-    ):
-        initial_state = State()
-        self.history_db = history_db  # maybe put this into `state` ?
-        super().__init__(name, redis_url, channels, initial_state)
-
-    async def message_handler(
-        self, channel: str, data: str, pattern: str, state: State
-    ) -> None:
-        message = json.loads(data)
-        # log.info("channel: %s, pattern: %s", channel, pattern)
-
-        if channel == "coordinate":
-            await self.coordinate_handler(message, state)
-
-        if channel == "altitude":
-            await self.altitude_handler(message, state)
-
-    async def coordinate_handler(self, message: dict[str, Any], state: State) -> None:
-        ts: TimeSeries = self.redis.ts()  # type: ignore
-        retention_msecs = 1000 * 60
-
-        icao24 = message["icao24"]
-        timestamp_ms = int(float(message["timestamp"]) * 1000)
-        latitude, longitude = float(message["latitude"]), float(message["longitude"])
-        record = {
-            "icao24": icao24,
-            "last": timestamp_ms,
-            "latitude": latitude,
-            "longitude": longitude,
-            "altitude": None,  # no altitude from coordinate message
-        }
-        self.history_db.insert_tracks([record])
-        log.debug("persiste record in db for %s", icao24)
-
-        # EXPERIMENTAL: store latitude and longitude in redis timeseries
-        latitude_key, longitude_key = f"latitude:{icao24}", f"longitude:{icao24}"
-        labels = {
-            "type": "latlong",
-            "icao24": icao24,
-        }
-        if not await self.redis.exists(latitude_key):
-            result = await ts.create(
-                latitude_key, retention_msecs=retention_msecs, labels=labels
-            )
-            log.info("add latitude_key %s %s", latitude_key, result)
-
-        if not await self.redis.exists(longitude_key):
-            result = await ts.create(
-                longitude_key, retention_msecs=retention_msecs, labels=labels
-            )
-            log.info("add longitude_key %s %s", longitude_key, result)
-
-        fields = ["timestamp", "icao24", "latitude", "longitude"]
-        message = {k: message[k] for k in fields}
-        values = [
-            (latitude_key, timestamp_ms, latitude),
-            (longitude_key, timestamp_ms, longitude),
-        ]
-        await ts.madd(values)  # type: ignore
-
-    async def altitude_handler(self, message: dict[str, Any], state: State) -> None:
-        ts: TimeSeries = self.redis.ts()  # type: ignore
-        retention_msecs = 1000 * 60
-
-        icao24 = message["icao24"]
-        timestamp_ms = int(float(message["timestamp"]) * 1000)
-
-        if message["altitude"] is None:
-            return
-
-        altitude = float(message["altitude"])
-
-        altitude_key = f"altitude:{icao24}"
-        labels = {"type": "altitude", "icao24": icao24}
-        if not await self.redis.exists(altitude_key):
[…]
