+++ packages/tangram_example/src/tangram_example/index.ts
@@ -1,0 +1,6 @@
+import type { TangramApi } from "@open-aviation/tangram/types";
+import ExampleWidget from "./ExampleWidget.vue";
+
+export function install(api: TangramApi) {
+  api.registerWidget("example", ExampleWidget);
+}
