### 2025-09-05 5ab56b7
docs: fix stupid yaml syntax preventing docs from building without installation
