### 2025-09-06 #8662 @xoolive
Thank you, @fredshone, for the useful feedback.
(Been very busy these days too, no worry!)

- I added a line in the quickstart with the `podman system prune` [here](https://mode-s.org/tangram/quickstart/#troubleshooting)
- I added a section for plane [here](https://mode-s.org/tangram/architecture/planes/#testing), trajectory [here](https://mode-s.org/tangram/architecture/trajectory/#testing) and channels [here](https://mode-s.org/tangram/architecture/channel/#testing) about testing. I agree that faking an interface as you say is the way to go. We are considering to add some replay functionalities in the future, which would be a perfect opportunity to present these interfaces and include better testing.

Please be sure we are clearly aware that the state of the testing is not ideal at the moment, and that improving it will be a priority in future versions.
