### 2025-09-10 22cfe58
refactor: systeminfo (#38)
this is a first step in a larger migration of the entire frontend.

to focus this change solely on logic, all v0.1 html and css for migrated
components have been copied verbatim.

- renders widgets into designated locations (`TopBar`, `SideBar`)
- initial implementation `UiApi` for widget registration and the
  `RealtimeApi` for WebSocket subscriptions
- use Vue's `inject` mechanism for subscribing to real-time updates.
- a visual redesign is out of scope for this change: while the use of bootstrap
  is extremely questionable, we largely keep it.
