+++ packages/tangram/src/tangram/main.ts
@@ -1,4 +1,5 @@
 import { createApp } from "vue";
 import App from "./App.vue";
+import "./user.css";
 
 createApp(App).mount("#app");
