+++ packages/tangram/src/tangram/api.ts
@@ -1,0 +1,290 @@
+import {
+  App,
+  ref,
+  Component,
+  computed,
+  reactive,
+  type Ref,
+  type ComputedRef
+} from "vue";
+import L from "leaflet";
+import type { LatLngBounds, Layer, Map as LeafletMap } from "leaflet";
+import { Socket, Channel } from "phoenix";
+
+class NotImplementedError extends Error {
+  constructor(message = "this function is not yet implemented.") {
+    super(message);
+    this.name = "NotImplementedError";
+  }
+}
+
+export interface ChannelConfig {
+  host: string;
+  port: number;
+}
+
+export interface TangramConfig {
+  channel: ChannelConfig;
+}
+
+export type EntityId = string;
+
+export interface Disposable {
+  dispose(): void;
+}
+
+export interface Entity<TState extends EntityState = EntityState> {
+  readonly id: EntityId;
+  readonly type: string;
+  readonly state: TState;
+}
+
+export type EntityState = unknown;
+
+export interface IPosition {
+  readonly lat: number;
+  readonly lng: number;
+}
+
+export interface ITimestamp {
+  readonly timestamp: Date;
+}
+
+export type WidgetLocation = "TopBar" | "SideBar" | "MapOverlay";
+
+/* Local time. Not to be confused with the server time. */
+export class TimeApi implements Disposable {
+  readonly now = ref(new Date());
+  readonly isPlaying = ref(false);
+
+  constructor() {}
+
+  dispose() {}
+
+  play(): void {
+    throw new NotImplementedError();
+  }
+  pause(): void {
+    throw new NotImplementedError();
+  }
+  seek(time: Date): void {
+    throw new NotImplementedError();
+  }
+}
+
+export class UiApi {
+  private app: App;
+  readonly widgets = reactive<Record<WidgetLocation, { id: string }[]>>({
+    TopBar: [],
+    SideBar: [],
+    MapOverlay: []
+  });
+
+  constructor(app: App) {
+    this.app = app;
+  }
+
+  registerWidget(
+    id: string,
+    location: WidgetLocation,
+    component: Component
+  ): Disposable {
+    this.app.component(id, component);
+
+    const widget = { id };
+    this.widgets[location].push(widget);
+
+    return {
+      dispose: () => {
+        const locationWidgets = this.widgets[location];
+        const index = locationWidgets.findIndex(w => w.id === id);
+        if (index > -1) {
+          locationWidgets.splice(index, 1);
+        }
+      }
+    };
+  }
+}
+
+export class MapApi {
+  readonly bounds: Ref<Readonly<LatLngBounds>> = ref(
+    new L.LatLngBounds([0, 0], [0, 0])
+  );
+
+  setView(position: { lat: number; lng: number }, zoom: number): void {
+    throw new NotImplementedError();
+  }
+  addLayer(layer: Layer): Disposable {
+    throw new NotImplementedError();
+  }
+  getMapInstance(): LeafletMap {
+    throw new NotImplementedError();
+  }
+}
+
+export class StateApi {
+  readonly entities: Ref<ReadonlyMap<EntityId, Entity>> = ref(new Map());
+  readonly activeEntityId: Ref<EntityId | null> = ref(null);
+  readonly activeEntity: ComputedRef<Entity | null> = computed(() => null);
+
+  registerEntityType(type: string): void {
+    throw new NotImplementedError();
+  }
+  getEntitiesByType<T extends EntityState>(
+    type: string
+  ): ComputedRef<ReadonlyMap<EntityId, Entity<T>>> {
+    throw new NotImplementedError();
+  }
+  setActiveEntity(entityId: EntityId): void {
+    throw new NotImplementedError();
+  }
[…]
