+++ packages/tangram_jet1090/src/tangram_jet1090/index.ts
@@ -1,6 +1,6 @@
-import type { TangramApi } from "@open-aviation/tangram/types";
+import type { TangramApi } from "@open-aviation/tangram/api";
 import Jet1090Widget from "./Jet1090Widget.vue";
 
 export function install(api: TangramApi) {
-  api.registerWidget("jet1090", Jet1090Widget);
+  api.ui.registerWidget("jet1090-widget", "SideBar", Jet1090Widget);
 }
