+++ packages/tangram_example/src/tangram_example/index.ts
@@ -1,6 +1,6 @@
-import type { TangramApi } from "@open-aviation/tangram/types";
+import type { TangramApi } from "@open-aviation/tangram/api";
 import ExampleWidget from "./ExampleWidget.vue";
 
 export function install(api: TangramApi) {
-  api.registerWidget("example", ExampleWidget);
+  api.ui.registerWidget("example-widget", "SideBar", ExampleWidget);
 }
