+++ packages/tangram/src/tangram/plugin.ts
@@ -1,13 +1,19 @@
-import type { TangramApi } from "./types";
+import type { TangramApi } from "./api";
 
 export async function loadPlugins(tangramApi: TangramApi) {
   const manifest = await fetch("/manifest.json").then(res => res.json());
 
   for (const pluginName of manifest.plugins) {
     const pluginManifestUrl = `/plugins/${pluginName}/plugin.json`;
-    const pluginManifest = await fetch(pluginManifestUrl).then(res =>
-      res.json()
-    );
+    const pluginManifest = await fetch(pluginManifestUrl).then(res => res.json());
+
+    if (pluginManifest.style) {
+      const link = document.createElement("link");
+      link.rel = "stylesheet";
+      link.href = `/plugins/${pluginName}/${pluginManifest.style}`;
+      document.head.appendChild(link);
+    }
+
     const entryPointUrl = `/plugins/${pluginName}/${pluginManifest.main}`;
 
     try {
@@ -16,7 +22,7 @@ export async function loadPlugins(tangramApi: TangramApi) {
         pluginModule.install(tangramApi);
       }
     } catch (e) {
-      console.error(`Failed to load plugin "${pluginName}":`, e);
+      console.error(`failed to load plugin "${pluginName}":`, e);
     }
   }
 }
