+++ eslint.config.ts
@@ -1,0 +1,18 @@
+import { globalIgnores } from "eslint/config";
+import {
+  defineConfigWithVueTs,
+  vueTsConfigs
+} from "@vue/eslint-config-typescript";
+import pluginVue from "eslint-plugin-vue";
+import skipFormatting from "@vue/eslint-config-prettier/skip-formatting";
+
+export default defineConfigWithVueTs(
+  {
+    name: "app/files-to-lint",
+    files: ["**/*.{ts,mts,js,mjs,vue}"]
+  },
+  globalIgnores(["**/dist/**", "**/dist-frontend/**", "packages/_jet1090_legacy/**", "web_legacy/**", ".venv"]),
+  pluginVue.configs["flat/recommended"],
+  vueTsConfigs.recommended,
+  skipFormatting  // use prettier for code formatting, eslint for code quality
+);
