### 2025-09-10 f5abcc8
format: setup eslint, fix clippy
