### 2025-09-10 cbd2269
format: setup eslint, fix clippy
