### 2025-09-10 63dbfa6
format: setup eslint, fix clippy
