+++ tangram.example.toml
@@ -1,7 +1,7 @@
 [core]
 redis_url = "redis://localhost:6379"
 plugins = [
-    "tangram_example",
+    # "tangram_example",
     "tangram_jet1090",
     "tangram_system",
     "tangram_weather"
