+++ packages/tangram_jet1090/src/tangram_jet1090/PlanePath.ts
@@ -1,0 +1,806 @@
+export interface IconProps {
+  x: number;
+  y: number;
+  scale: number;
+  rotcorr: number;
+  path: string;
+  ofX?: number;
+  ofY?: number;
+}
+
+export interface AcSymbol {
+  id: string;
+  sym: string;
+}
+
+let acImagePathsR: Record<string, IconProps> = {
+  // triangle
+  $0: { x: 12, y: 12, scale: 1, rotcorr: 180, path: "M0,0 6,12 12,0z" },
+  // small aircraft
+  $1: {
+    x: 16,
+    y: 17,
+    scale: 1,
+    rotcorr: 180,
+    path: "M8,17 9,16 9,12 16,11 16,8 9,7 9,3 11,1 11,0 8,1 5,0 5,1 7,3 7,7 0,8 0,11 7,12 7,16 8,17z"
+  },
+  // two propeller aircraft
+  $2: {
+    x: 26,
+    y: 18,
+    scale: 1,
+    rotcorr: 180,
+    path: "M13,18 14,17 14,17 15,14 17,14 17,16 18,17 19,16 19,14 25,14 26,13 26,12 26,12 25,11 19,10 19,9 18,9 18,10 15,9 14,6 14,2 18,1 18,0 13,0 8,0 8,1 12,2 12,6 11,9 8,10 8,9 7,9 7,10 1,11 0,12 0,12 0,13 1,14 7,14 7,16 8,17 9,16 9,14 11,14 12,17 12,17 13,18z"
+  },
+  // jet with two turbines at back (Learjet or so)
+  $3: {
+    x: 21,
+    y: 24,
+    scale: 1,
+    rotcorr: 180,
+    path: "M11,24 12,23 12,14 20,10 21,8 13,10 12,9 12,7 14,7 13,4 12,4 12,3 14,2 15,0 12,1 11,0 10,1 7,0 8,2 10,3 10,4 9,4 8,7 10,7 10,9 9,10 0,8 1,10 10,14 10,23 11,24z"
+  },
+  // A320, B737
+  $4: {
+    x: 24,
+    y: 28,
+    scale: 1,
+    rotcorr: 180,
+    path: "M12,28 13,26 14,23 14,18 16,17 16,19 18,19 18,16 24,11 24,10 23,10 15,13 14,13 13,6 13,5 16,2 16,0 13,2 12,0 11,2 8,0 8,2 11,5 11,6 10,13 9,13 1,10 0,10 0,11 6,16 6,19 8,19 8,17 10,18 10,23 11,26 12,28z"
+  },
+  // B747, A380
+  $5: {
+    x: 30,
+    y: 31,
+    scale: 1,
+    rotcorr: 180,
+    path: "M15,31 17,28 17,22 20,20 20,22 22,22 22,19 25,17 25,19 27,19 27,16 30,13 30,11 19,14 17,14 17,7 16,5 20,2 20,0 16,1 15,0 14,1 10,0 10,2 14,5 13,7 13,14 11,14 0,11 0,13 3,16 3,19 5,19 5,17 8,19 8,22 10,22 10,20 13,22 13,28 15,31z"
+  },
+  // Fighter
+  $6: {
+    x: 16,
+    y: 20,
+    scale: 1,
+    rotcorr: 180,
[…]
