### 2025-10-02 321f2c0
feat: aircraft info widget, charts, trail, city pair, shared httpx client
added rs1090 for aircraft information.
note that it has been added to the core instead of the jet1090
plugin because of an annoying vite bug (see `tangram/vite.config.ts` for more
information) https://github.com/vitejs/vite/issues/4454

fix: aircraft count not reactive
