+++ packages/tangram_jet1090/src/tangram_jet1090/index.ts
@@ -2,6 +2,9 @@ import { watch } from "vue";
 import type { TangramApi, Entity } from "@open-aviation/tangram/api";
 import AircraftLayer from "./AircraftLayer.vue";
 import AircraftCountWidget from "./AircraftCountWidget.vue";
+import AircraftInfoWidget from "./AircraftInfoWidget.vue";
+import AircraftTrailLayer from "./AircraftTrailLayer.vue";
+import init, { run } from "rs1090-wasm";
 
 interface RawAircraft {
   icao24: string;
@@ -9,27 +12,34 @@ interface RawAircraft {
 }
 
 export function install(api: TangramApi) {
+  // initialising webassembly manually because esm build isn't supported well
+  (async () => {
+    await init("/rs1090_wasm_bg.wasm");
+    run();
+  })();
   api.ui.registerWidget("jet1090-count-widget", "TopBar", AircraftCountWidget);
   api.ui.registerWidget("jet1090-aircraft-layer", "MapOverlay", AircraftLayer);
+  api.ui.registerWidget("jet1090-info-widget", "SideBar", AircraftInfoWidget);
+  api.ui.registerWidget("jet1090-trail-layer", "MapOverlay", AircraftTrailLayer);
 
   api.state.registerEntityType("aircraft");
 
-  const stopWatch = watch(
-    api.realtime.connectionId,
-    newId => {
-      if (newId) {
-        subscribeToAircraftData(api, newId);
-        stopWatch();
-      }
-    },
-    { immediate: true }
-  );
+  (async () => {
+    try {
+      const connectionId = await api.realtime.ensureConnected();
+      await subscribeToAircraftData(api, connectionId);
+    } catch (e) {
+      console.error("failed initializing jet1090 realtime subscription", e);
+    }
+  })();
+
   watch(
     api.map.bounds,
     newBounds => {
-      if (!api.realtime.connectionId.value) return;
+      const connId = api.realtime.getConnectionId();
+      if (!connId) return;
       const payload = {
-        connectionId: api.realtime.connectionId.value,
+        connectionId: connId,
         northEastLat: newBounds.getNorthEast().lat,
         northEastLng: newBounds.getNorthEast().lng,
         southWestLat: newBounds.getSouthWest().lat,
