+++ packages/tangram/pyproject.toml
@@ -19,7 +19,7 @@ dependencies = [
     "fastapi>=0.115.4",
     "uvicorn>=0.32.0",
     "redis>=5.2.0",
-    "httpx>=0.27.2",
+    "httpx[http2]>=0.27.2",
     "pydantic>=2.6.1",
     "typer",                          # provides rich
     "tomli; python_version < '3.11'",
