### 2025-10-02 d9f7363
refactor: airport search plugin
this is a frontend-only plugin that relies of `rs1090` in the core.
