+++ packages/tangram_airports/pyproject.toml
@@ -1,0 +1,24 @@
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[project]
+name = "tangram_airports"
+version = "0.2.0"
+dependencies = ["tangram>=0.2.0"]
+authors = [
+    { name = "Xavier Olive", email = "git@xoolive.org" },
+]
+
+[project.entry-points."tangram.plugins"]
+tangram_airports = "tangram_airports:plugin"
+
+[tool.uv.sources]
+tangram = { workspace = true }
+
+[tool.hatch.build.targets.sdist]
+ignore-vcs = true
+include = ["dist-frontend/*", "src/*"]
+
+[tool.hatch.build.targets.wheel.force-include]
+"dist-frontend" = "tangram_airports/dist-frontend"
