### 2025-10-02 69bed4c
refactor: airport search plugin
this is a frontend-only plugin that relies of `rs1090` in the core.
