+++ packages/tangram_airports/src/tangram_airports/__init__.py
@@ -1,0 +1,5 @@
+import tangram
+
+plugin = tangram.Plugin(
+    frontend_path="dist-frontend",
+)
