+++ packages/tangram_airports/vite.config.ts
@@ -1,0 +1,6 @@
+import { defineConfig } from "vite";
+import { tangramPlugin } from "@open-aviation/tangram/vite-plugin";
+
+export default defineConfig({
+  plugins: [tangramPlugin()],
+});
