+++ tangram.example.toml
@@ -4,6 +4,7 @@ plugins = [
     # "tangram_example",
     "tangram_jet1090",
     "tangram_system",
+    "tangram_airports",
     "tangram_weather"
 ]
 
