+++ packages/tangram_airports/vite.config.ts
@@ -2,5 +2,5 @@ import { defineConfig } from "vite";
 import { tangramPlugin } from "@open-aviation/tangram/vite-plugin";
 
 export default defineConfig({
-  plugins: [tangramPlugin()],
+  plugins: [tangramPlugin()]
 });
