+++ docs/plugins/examples/citypair.md
@@ -1,168 +1,23 @@
-# Origin and destination city pair
-
-!!! danger "Legacy Frontend"
-    This example describes the plugin system for the legacy v0.1 frontend, which is configured via `.env` and `web_legacy/vite.config.js`. The following guide is scheduled for an update.
+# Origin and Destination
 
 ## Statement of need
 
-The city pair plugin displays the origin and destination information for the selected aircraft, showing both the airport codes (ICAO) and the city names. This makes it easy to quickly understand the flight route without leaving the main map interface.
+The `tangram_jet1090` plugin includes a widget that displays the origin and destination information for a selected aircraft, showing both the airport ICAO codes and the city names. This makes it easy to quickly understand the flight's route without leaving the main map interface.
 
 ## Implementation
 
-The implementation of the city pair plugin is a Vue.js component that displays the origin and destination of a flight when an aircraft is selected on the map. The implementation is located in the `web_legacy/src/plugins/CityPair.vue` file.
+The city pair widget is part of the `AircraftInfoWidget.vue` component, which appears in the sidebar when an aircraft is selected.
 
-!!! tip
+It works as follows:
 
-    When a Vue component is a plugin, its implementation can be overridden by a custom implementation located in a different directory. This is useful if you want to have several possible implementations for a functionality, or if you want to use a different implementation for a specific use case.
-    This is particularly useful if you have a custom implementation of the city pair plugin using different sources of data (specific airport or airline development API, EUROCONTROL B2B, etc.) that are not necessarily publicly available, or not without a private token.
+1. When an aircraft is selected, the widget uses the aircraft's callsign to make a request to the `/route/{callsign}` API endpoint.
+2. This endpoint, provided by the `tangram_jet1090` backend plugin, proxies the request to the [OpenSky Network's route database](https://flightroutes.opensky-network.org).
+3. If a route is found, the widget displays the origin and destination airport ICAO codes.
+4. It then uses the [`rs1090-wasm` library](https://www.npmjs.com/package/rs1090-wasm) (bundled with `tangram` core) to look up and display the corresponding city names for each airport.
 
-    If the environment variable `TANGRAM_WEB_CITYPAIR_PLUGIN` is defined, the plugin will be loaded from the specified path. Otherwise, it will be loaded from the default fallback location. You can specify the path to your custom implementation in the `.env` file.
-
-In this example, the flight route information is obtained from the OpenSky Network's route service, which provides origin-destination pairs based on the flight's callsign.
-The city names are retrieved using the `airport_information` function from the [`rs1090-wasm` library](https://www.npmjs.com/package/rs1090-wasm).
-
-### 1. Declare the plugin in the `vite.config.js` file
-
-In order to use the `CityPair` component as a plugin, you need to declare it in the `vite.config.js` file:
-
-```javascript
-plugins: [
-  // ..., other settings
-  dynamicComponentsPlugin({
-    envPath: "../.env",
-    fallbackDir: "/src/plugins/",
-    availablePlugins: [
-      "airportSearch",
-      "systemInfo",
-      "sensorsInfo",
-      // "windField",
-      "cityPair",  // <-- new line
-    ],
-  }),
-],
-```
-
-### 2. Configure proxy for the route API
-
-The city pair plugin needs to fetch route information from an external API.
-
-When we fetch data from an external API, we often encounter CORS (Cross-Origin Resource Sharing) issues, which is a security feature implemented by web browsers to prevent malicious websites from making requests to other domains. This can block our requests to the OpenSky Network's route service.
[…]

[…]
