+++ eslint.config.ts
@@ -11,7 +11,7 @@ export default defineConfigWithVueTs(
     name: "app/files-to-lint",
     files: ["**/*.{ts,mts,js,mjs,vue}"]
   },
-  globalIgnores(["**/dist/**", "**/dist-frontend/**", "packages/_jet1090_legacy/**", "web_legacy/**", ".venv"]),
+  globalIgnores(["**/dist/**", "**/dist-frontend/**", "packages/_jet1090_legacy/**", ".venv"]),
   pluginVue.configs["flat/recommended"],
   vueTsConfigs.recommended,
   skipFormatting  // use prettier for code formatting, eslint for code quality
