+++ docs/plugins/airports.md
@@ -13,10 +13,3 @@ This is a frontend-only plugin that requires no additional backend configuration
 ## Configuration
 
 To enable this plugin, add `"tangram_airports"` to the `plugins` list in your `tangram.toml` file:
-
-```toml
-[core]
-plugins = [
-    "tangram_airports",
-    # ... other plugins
-]

[…]
