### 2025-10-03 97b1f1b
refactor: sensors, weather windfield
docs: removed outdated vue2 examples for now. in the future, we can use
`pymdownx.snippets` to avoid having to rewrite everything.

the unfinished slider component was moved into `_jet1090_legacy` for now.
