### 2025-10-03 197b253
feat: show plugin loading progress
