### 2025-10-04 a37a940
local fixes
