+++ src/tangram/app.py
@@ -2,10 +2,10 @@ from __future__ import annotations
 
 import json
 import logging
+import pathlib
 import uuid
 from datetime import datetime
 from typing import Any, Dict
-import pathlib
 
 from fastapi import FastAPI, Request, WebSocket, status
 from fastapi.concurrency import run_until_first_complete
@@ -16,6 +16,7 @@ from starlette.responses import HTMLResponse
 
 from tangram import websocket as tangram_websocket
 from tangram.plugins import rs1090_source
+from tangram.util.geojson import geojson_fp
 
 log = logging.getLogger("tangram")
 
@@ -60,7 +61,9 @@ app = FastAPI(
     ],
 )
 
-app.mount("/static", StaticFiles(directory=tangram_module_root / "static"), name="static")
+app.mount(
+    "/static", StaticFiles(directory=tangram_module_root / "static"), name="static"
+)
 app.mount("/plugins/rs1090", rs1090_source.rs1090_app, name="rs1090")
 
 start_time = datetime.now()
@@ -110,9 +113,16 @@ async def trajectory(icao24: str) -> Dict[str, Any]:
     }
     return geojson
 
+
+@app.get("/fp.geojson")
+def fp():
+    return geojson_fp()
+
+
 @app.get("/data/{icao24}")
 async def data(icao24: str) -> list[dict[str, Any]]:
     return await rs1090_source.icao24_track(rs1090_source.BASE_URL + "/track", icao24)
+
 
 @app.get("/turb.geojson")
 async def turbulence() -> Dict[str, Any]:
