### 2025-10-05 7a6f13f
fix: get rid of bootstrap, replace font-awesome cdn with local
this allows the frontend to load without an internet connection
