### 2025-10-05 0f1d7ec
fix: get rid of bootstrap, replace font-awesome cdn with local
this allows the frontend to load without an internet connection
