### 2025-10-20 f1c5294
updates to paper and bibliography
