+++ paper/paper.bib
@@ -4,6 +4,27 @@
   url    = {https://www.eurocontrol.int/dashboard/rnd-data-archive}
 }
 
+@misc{channel,
+  title  = {{WebSocket}-based {Real}-{Time Data Streaming Framework}},
+  author = {Huang, Xiaogang},
+  url    = {https://github.com/emctoo/channel},
+  year   = {2025}
+}
+
+@misc{jet1090,
+  title  = {jet1090: a {Real}-{Time} {Mode S and ADS-B Decoder} in {Rust}},
+  author = {Olive, Xavier},
+  url    = {https://mode-s.org/jet1090/},
+  year   = {2025}
+}
+
+@misc{dump1090,
+  title  = {dump1090: a simple {Mode}-{S} {Decoder for {RTLSDR} devices}},
+  author = {Sanfilippo, Salvatore},
+  url    = {https://github.com/antirez/dump1090},
+  year   = {2013}
+}
+
 @article{olive_traffic_2019,
   title   = {{traffic}, a Toolbox for Processing and Analysing Air Traffic Data},
   author  = {Olive, Xavier},
@@ -90,3 +111,33 @@
   doi     = {10.33012/navi.657},
   url     = {https://navi.ion.org/content/71/3/navi.657}
 }
+
+@article{basora_recent_2019,
+  title   = {Recent {{Advances}} in {{Anomaly Detection Methods Applied}} to {{Aviation}}},
+  author  = {Basora, Luis and Olive, Xavier and Dubot, Thomas},
+  year    = {2019},
+  month   = oct,
+  journal = {Aerospace},
+  volume  = {6},
+  number  = {11},
+  pages   = {117},
+  doi     = {10.3390/aerospace6110117},
+  url     = {https://www.mdpi.com/2226-4310/6/11/117}
+}
+
+
+
+@misc{jarry_neural_2025,
+  title         = {A {{Neural ODE Approach}} to {{Aircraft Flight Dynamics Modelling}}},
+  author        = {Jarry, Gabriel and Dalmau, Ramon and Olive, Xavier and Very, Philippe},
+  year          = {2025},
+  month         = sep,
+  number        = {arXiv:2509.23307},
+  eprint        = {2509.23307},
+  publisher     = {arXiv},
+  doi           = {10.48550/arXiv.2509.23307},
+  url           = {http://arxiv.org/abs/2509.23307},
+  archiveprefix = {arXiv},
+  annotation    = {preprint}
+}
+
