### 2025-10-20 #8662 @anastassiavybornova
Hi again! @xoolive looking great, there is only one open to-do from the list above, please rephrase (or just remove entirely) the "published by the authors in the literature" part in the potential applications section
