### 2025-10-20 d7a8c7c
Rephrase awkward sentence
