### 2025-10-20 d813848
fix references
