+++ paper/paper.bib
@@ -4,21 +4,21 @@
   url    = {https://www.eurocontrol.int/dashboard/rnd-data-archive}
 }
 
-@misc{channel,
+@misc{channel_github,
   title  = {{WebSocket}-based {Real}-{Time Data Streaming Framework}},
   author = {Huang, Xiaogang},
   url    = {https://github.com/emctoo/channel},
   year   = {2025}
 }
 
-@misc{jet1090,
+@misc{jet1090_github,
   title  = {jet1090: a {Real}-{Time} {Mode S and ADS-B Decoder} in {Rust}},
   author = {Olive, Xavier},
   url    = {https://mode-s.org/jet1090/},
   year   = {2025}
 }
 
-@misc{dump1090,
+@misc{dump1090_github,
   title  = {dump1090: a simple {Mode}-{S} {Decoder for {RTLSDR} devices}},
   author = {Sanfilippo, Salvatore},
   url    = {https://github.com/antirez/dump1090},
