### 2025-10-30 #8662 @xoolive
Thank you @samhforbes

EUROCONTROL is the name of a European agency and is typeset with all caps.
https://www.eurocontrol.int/

About the other papers, we kept the capitalization intact as delivered by the website referenced by the DOI.

About the papers you mention, the following websites (you can as well click on the DOI in the pdf) come with a given capitalization that has been copied as is:
- https://journals.open.tudelft.nl/joas/article/view/7943
- https://www.sciencedirect.com/science/article/pii/S136192092030715X

Do you want us to override the official capitalization?
