### 2025-11-01 1b0b391
Add citation information in citation.cff
